graph/iterator: reflect the swiss map iterator shape
